requirements-web.txt
# Dashboard web app dependencies (used by Dockerfile.web)
streamlit>=1.35.0
pandas>=2.1.0
numpy>=1.24.0
plotly>=5.18.0
azure-storage-blob>=12.19.0
azure-storage-file-datalake>=12.14.0
azure-identity>=1.15.0
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0
